Financial - Move Payment::create idempotency tests to the APIv4 suite
They exercise the APIv4 Payment.create action, so they belong under tests/phpunit/api/v4 rather than the v3 PaymentTest.
